% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{as.data.frame.table1}
\alias{as.data.frame.table1}
\title{Convert a \code{table1} object to a \code{data.frame}.}
\usage{
\method{as.data.frame}{table1}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{table1}}.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame}.
}
\description{
Convert a \code{table1} object to a \code{data.frame}.
}
