% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{column_label}
\alias{column_label}
\title{A function to create column names with sample sizes}
\usage{
column_label(x)
}
\arguments{
\item{Grouping}{variable}
}
\description{
A function to create column names with sample sizes
}
\examples{
column_label()
}
\keyword{Column}
\keyword{labels}
\keyword{sample}
\keyword{size}
\keyword{with}
