% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{dist_check}
\alias{dist_check}
\title{A function to describe the distribution of a variable 
Tabulates and displays mean, geometric mean, median, proportion missing, 
histogram and histogram of the log outcome}
\usage{
dist_check(var1)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
A function to describe the distribution of a variable 
Tabulates and displays mean, geometric mean, median, proportion missing, 
histogram and histogram of the log outcome
}
\examples{
dist_check()
}
\keyword{distribution}
\keyword{mean,}
\keyword{median,}
