% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{factorp}
\alias{factorp}
\title{Factor creation function that preserves column attributes.}
\usage{
factorp(x, ...)
}
\arguments{
\item{x}{An object to be converted to a a \code{\link{factor}}.}

\item{...}{Further arguments passed to \code{\link{factor}}.}
}
\value{
An \code{factor} with the attributes of \code{x} other than
\code{"class"} and \code{"levels"} preserved.
}
\description{
Factor creation function that preserves column attributes.
}
\seealso{
\code{\link{factor}}
}
\keyword{utilities}
