% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{final_table}
\alias{final_table}
\title{A function to create the final table}
\usage{
final_table(var, group)
}
\arguments{
\item{Variables}{of interest and group variable}
}
\description{
A function to create the final table
}
\examples{
final_table()
}
\keyword{1}
\keyword{Table}

