% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{final_tableb}
\alias{final_tableb}
\title{A function to create the final table, with row percentages}
\usage{
final_tableb(var, group)
}
\arguments{
\item{Variables}{of interest and group variable}
}
\description{
A function to create the final table, with row percentages
}
\examples{
final_tableb()
}
\keyword{1}
\keyword{Table}

