% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{fx_model_linear}
\alias{fx_model_linear}
\title{A function to create a summary of a linear regression model.}
\usage{
fx_model_linear(y, x, data)
}
\arguments{
\item{y}{outcome variable of interest to be pasted in formula, in quotation marks.}

\item{x}{a single explanatory variable, or multiple variables separated by +, in quotation marks.}

\item{data}{a data frame containing the variables listed in x and y.}
}
\description{
A function to create a summary of a linear regression model.
}
\details{
Table with estimates, 95% CIs and p-values.
}
\examples{
iris$sepal_di<-as.factor((iris$Sepal.Length<5)*1)
#set factor levels
levels(iris$sepal_di)=c("<5",'>=5')
levels(iris$Species)=c('Setosa','Versicolor','Virginica')
#Set Labels
label(iris$Species)='Species'
label(iris$sepal_di)='Sepal Length'
label(iris$Sepal.Length)="Sepal Length"
label(iris$Sepal.Width)='Sepal Width'
label(iris$Petal.Length)='Petal Length'
label(iris$Petal.Width)='Petal Width'
mod_tab<-fx_model_linear('Sepal.Length',c('Species','Petal.Length'),iris)
}
\keyword{Linear}
\keyword{regression,}
\keyword{summary}
\keyword{table}
