% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{fx_rows}
\alias{fx_rows}
\title{A function to determine which row function to employ}
\usage{
fx_rows(variable, group, margin, single, ron, summary.stat)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
A function to determine which row function to employ
}
\examples{
fx_rows()
}
\keyword{Categorical,}
\keyword{continuous}
\keyword{variables}
