% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{fx_rowsb}
\alias{fx_rowsb}
\title{A function to determine which row function to employ, including row based percentages}
\usage{
fx_rowsb(var, group)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
A function to determine which row function to employ, including row based percentages
}
\examples{
fx_rowsb()
}
\keyword{Categorical,}
\keyword{continuous}
\keyword{variables}

