% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{fx_symbol}
\alias{fx_symbol}
\title{A function to generate a * to add to the label for non-normally distributed continuous variables}
\usage{
fx_symbol(y, x)
}
\arguments{
\item{Variable}{of interest}
}
\description{
A function to generate a * to add to the label for non-normally distributed continuous variables
}
\examples{
fx_symbol()
}
\keyword{*}
\keyword{indication}
