% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{geom_mean_row}
\alias{geom_mean_row}
\title{A row function}
\usage{
geom_mean_row(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates geometric mean and 95% CI, p-value from t-test or ANOVA
}
\examples{
geom_mean_row()
}
\keyword{Continuous}

