% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{ggplot_bar}
\alias{ggplot_bar}
\title{Title}
\usage{
ggplot_bar(
  dataset,
  group,
  VarName,
  Height = px(50),
  color_list = color_list_temp
)
}
\arguments{
\item{color_list}{}
}
\value{

}
\description{
Title
}
