% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{indices}
\alias{indices}
\title{Extract indices from and indexed object.}
\usage{
indices(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{indexed}}.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector of indices extracted from object \code{x}.
}
\description{
Extract indices from and indexed object.
}
