% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{label}
\alias{label}
\alias{label<-}
\alias{has.label}
\title{Label attribute.}
\usage{
label(x)

label(x) <- value

has.label(x)
}
\arguments{
\item{x}{An object.}

\item{value}{A \code{character} specifying the label.}
}
\description{
Label attribute.
}
\section{Functions}{
\itemize{
\item \code{label<-}: Set label attribute.

\item \code{has.label}: Check for label attribute.
}}

\examples{
x <- 1:10
label(x) <- "Foo"
has.label(x)
label(x)
}
\keyword{utilities}
