% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{mean_row}
\alias{mean_row}
\title{A row function}
\usage{
mean_row(y, x, ron)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates mean and SD, p-value from t-test or ANOVA
}
\examples{
mean_row()
}
\keyword{Continuous}
