% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTab.R
\name{meanTab}
\alias{meanTab}
\title{A function to create a table of means and SDs}
\usage{
meanTab(df, vars, group = NULL)
}
\arguments{
\item{df}{dataframe}

\item{vars}{list of continuous variables to calculate}

\item{group}{grouping categorical variable}
}
\description{
For continuous variables, this function produces a table with the mean, SD, and N missing of each variable for each group, and a p-value for the Wilcoxon rank-sum test comparing groups (if group is specified)
}
\examples{
data(airquality)
meanTab(airquality, c('Ozone', 'Temp', 'Wind'), group = 'Month')
}
\keyword{descriptive}
\keyword{statistics}
