% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{median_row}
\alias{median_row}
\title{A row function}
\usage{
median_row(y, x, ron)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates median and IQR, p-value from Wilcox rank-sum or Kruskal-Wallis
}
\examples{
median_row()
}
\keyword{Continuous,}
\keyword{non-normal}
