% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{missing_row}
\alias{missing_row}
\title{A function to quantify missing data}
\usage{
missing_row(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates no. and % of missing data
}
\examples{
missing_row()
}
\keyword{Continuous,}
\keyword{non-normal}
