% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{missing_table}
\alias{missing_table}
\title{A function to create a summary of missing data for 1 or more groups.}
\usage{
missing_table(data, variables, group, col.names = T)
}
\arguments{
\item{data}{a data frame containing the variables listed in variables 
and the group variable.}

\item{variables}{a vector or scalar of variables of interest, in quotation marks.}

\item{group}{an R object. Grouping variable can have >=1 level(s).}

\item{col.names}{a logical value indicating whether the column labels with group name and sample size are desired.}
}
\description{
A function to create a summary of missing data for 1 or more groups.
}
\details{
Can handle >=1 group; number of columns will adjust automatically.
}
\examples{
iris$sepal_di<-as.factor((iris$Sepal.Length<5)*1)
#set factor levels
levels(iris$sepal_di)=c("<5",'>=5')
levels(iris$Species)=c('Setosa','Versicolor','Virginica')
#Set Labels
label(iris$sepal_di)='Sepal Length'
label(iris$Sepal.Length)="Sepal Length"
label(iris$Sepal.Width)='Sepal weight'
label(iris$Petal.Length)='Petal Length'
label(iris$Petal.Width)='Petal Width'
tab_miss<-missing_table(iris,c('Sepal.Length','Petal.Length','Sepal.Width','Petal.Width','sepal_di'),iris$Species)
}
\keyword{Missing}
\keyword{data,}
\keyword{summary}
\keyword{table}
