% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{print.table1}
\alias{print.table1}
\title{Print \code{table1} object.}
\usage{
\method{print}{table1}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{table1}}.}

\item{...}{Further arguments passed on to other \code{print} methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print \code{table1} object.
}
\details{
In an interactive context, the rendered table will be displayed in
a web browser. Otherwise, the HTML code will be printed as text.
}
