% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Row_functions.R
\name{prop_row}
\alias{prop_row}
\title{A row function}
\usage{
prop_row(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates no. and % for each level of the variable, p-value from Chi-sq test
}
\examples{
prop_row()

}
\keyword{>2}
\keyword{Categorical,}
\keyword{levels}

