% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{prop_row_one}
\alias{prop_row_one}
\title{A row function
This function creates no. and % for the second level of the variable, p-value from Chi-sq test}
\usage{
prop_row_one(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
A row function
This function creates no. and % for the second level of the variable, p-value from Chi-sq test
}
\examples{
prop_row_one()
}
\keyword{<3}
\keyword{Categorical,}
\keyword{levels}

