% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{prop_row_oneb}
\alias{prop_row_oneb}
\title{This function creates no. and % for the second level of the variable, % based on row total, p-value from Chi-sq test}
\usage{
prop_row_oneb(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates no. and % for the second level of the variable, % based on row total, p-value from Chi-sq test
}
\examples{
prop_row_oneb()
}
\keyword{<3}
\keyword{Categorical,}
\keyword{levels}

