% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Row_functions.R
\name{prop_rowb}
\alias{prop_rowb}
\title{A row function}
\usage{
prop_rowb(y, x)
}
\arguments{
\item{Variable}{of interest and group variable}
}
\description{
This function creates no. and % for each level of the variable, % based on row total, p-value from Chi-sq test
}
\examples{
prop_rowb()
}
\keyword{>2}
\keyword{Categorical,}
\keyword{levels}

