% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{render.strat.default}
\alias{render.strat.default}
\title{Render strata labels for table output.}
\usage{
render.strat.default(label, n, transpose = F)
}
\arguments{
\item{label}{A \code{character} vector containing the labels.}

\item{n}{A \code{numeric} vector containing the sizes.}
}
\value{
A \code{character}, which may contain HTML markup.
}
\description{
Called from \code{\link{table1.formula}} to render strata labels
for displaying in the table.
}
\keyword{internal}
