% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{stats.apply.rounding}
\alias{stats.apply.rounding}
\title{Apply rounding to basic descriptive statistics.}
\usage{
stats.apply.rounding(x, digits = 3, digits.pct = 1)
}
\arguments{
\item{x}{A list, as returned by \code{\link{stats.default}}.}

\item{digits}{Number of significant digits.}

\item{digits.pct}{Number of digits after the decimal place for percentages.}
}
\value{
A list with the same number of elements as \code{x}.
}
\description{
Not all statistics should be rounded; in particular min, max and median should
retain their original precision. This function will apply rounding
selectively to a list of statistics as returned by
\code{\link{stats.default}}. The rounded values will be \code{character},
not \code{numeric}, and will have 0 padding to ensure consistent number of
significant digits. Percentages are rounded to a fixed number of decimal
places (default 1) rather than a specific number of significant digits.
}
\examples{
x <- exp(rnorm(100, 1, 1))
stats.default(x)
stats.apply.rounding(stats.default(x), digits=3)

}
\seealso{
\code{\link{signif_pad}}
\code{\link{stats.default}}
}
\keyword{utilities}

