% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{t1kable}
\alias{t1kable}
\title{Convert a \code{table1} object to \code{kabelExtra}.}
\usage{
t1kable(x, booktabs = TRUE, ..., format, align, bold.headers = TRUE)
}
\arguments{
\item{x}{An object returned by \code{\link{table1}}.}

\item{booktabs}{Passed to \code{kbl} (default \code{TRUE}).}

\item{...}{Other options passed to \code{kbl}.}

\item{format}{Passed to \code{kbl} (optional).}

\item{align}{Passed to \code{kbl} (optional). The default is to left align
the labels (first column) and center everything else.}

\item{bold.headers}{Should the column headers be bolded?}
}
\value{
A \code{kabelExtra} object.
}
\description{
Convert a \code{table1} object to \code{kabelExtra}.
}
\note{
The \code{kableExtra} package needs to be installed for this to work.
}
