% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table.rows}
\alias{table.rows}
\title{Convert to HTML table rows.}
\usage{
table.rows(x, row.labels = rownames(x), th = FALSE)
}
\arguments{
\item{x}{A vector or table-like structure (e.g. a \code{\link{data.frame}} or \code{\link{matrix}}).}

\item{row.labels}{Values for the first column, typically used to label the row, or \code{NULL} to omit.}

\item{th}{A logical. Should \code{th} tags be used rather than \code{td}?}
}
\value{
A \code{character} which contains an HTML table fragment.
}
\description{
Many functions exist in R to generate HTML tables.  This function is useful
for generating an HTML table fragment (rather than a whole table), which can
be embedded in a larger table structure. Row labels, if specified, have a
special HTML \code{class} \code{rowlabel}, which can be useful as a hook to
customize their appearance using CSS; the first row label has an additional
\code{class} \code{firstrowlabel}.
}
\examples{
x <- matrix(signif_pad(exp(rnorm(100, 1, 1))), 10, 10)
table.rows(x, NULL)
table.rows(x, LETTERS[1:10])
table.rows(LETTERS[1:3], "Headings", th=TRUE)
}
\keyword{utilities}

