\name{Table1-package}
\alias{Table1-package}
\alias{Table1}
\docType{package}
\title{
\packageTitle{Table1}
}
\description{
\packageDescription{Table1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Table1}
\packageIndices{Table1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Table1}

Maintainer: \packageMaintainer{Table1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
