% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{units}
\alias{units}
\alias{units<-}
\alias{has.units}
\title{Units attribute.}
\usage{
units(x)

units(x) <- value

has.units(x)
}
\arguments{
\item{x}{An object.}

\item{value}{A \code{character} specifying the units}
}
\description{
Units attribute.
}
\section{Functions}{
\itemize{
\item \code{units<-}: Set units attribute.

\item \code{has.units}: Check for attribute.
}}

\examples{
x <- 1:10
units(x) <- "cm"
has.units(x)
units(x)
}
\keyword{utilities}
