% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{update_html}
\alias{update_html}
\title{Update HTML.}
\usage{
update_html(x)
}
\arguments{
\item{x}{An object returned by \code{\link{table1}}.}
}
\value{
An object of class "table1" which contains the updated HTML.
}
\description{
Used to (re-)generate the HTML code for a \code{link{table1}} object. In
most cases, this should not be used direction, unless you know what you are
doing.
}
