% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table1.R
\name{buildTable1Rows}
\alias{buildTable1Rows}
\title{Build rows for table-1.}
\usage{
buildTable1Rows(theVariable, theData, groupBy = NULL, percentFirst = TRUE,
  conductGroupTests = TRUE, stringsAsFactors = FALSE,
  meanDigits = getOption("digits"), sdDigits = getOption("digits"),
  freqDigits = getOption("digits"), statDigits = getOption("digits"),
  pDigits = getOption("digits"))
}
\arguments{
\item{theVariable}{A character vector specifying the name (column) to be processed}

\item{theData}{The data.frame containing the data for the variable.}

\item{groupBy}{A variable (factor) used to stratify the variable being reported (see details below).}

\item{percentFirst}{A boolean, if TRUE numeric variables are printed as x\%(n) otherwise as n(x\%).}

\item{conductGroupTests}{A boolean, if TRUE group comparisons are conducted (see details below).}

\item{stringsAsFactors}{A boolean, if TRUE, any variable that is of type character will get coerced to a factor.}

\item{meanDigits}{An integer indicating number of digits printed following the decimal place (this is not precision as in round, you get the number of digits you ask for, even if they are zero).}

\item{sdDigits}{An integer indicating number of digits printed following the decimal place for standard deviations (this is not precision as in round, you get the number of digits you ask for, even if they are zero).}

\item{freqDigits}{An integer indicating number of digits printed following the decimal place for percentages (this is not precision as in round, you get the number of digits you ask for, even if they are zero).}

\item{statDigits}{An integer indicating number of digits printed following the decimal place for test statistics (this is not precision as in round, you get the number of digits you ask for, even if they are zero).}

\item{pDigits}{An integer indicating number of digits printed following the decimal place for test P-Values (if numer of zeros in P is greater than pDigits the value will be "< 000...1").}
}
\value{
A data.frame.
}
\description{
One row is produced for numeric variables, multiple rows for factors.
}
\examples{
library(dplyr)
library(tidyr)
buildTable1Rows(theVariable = "age", theData = table1Dat, groupBy="site", percentFirst = TRUE,
conductGroupTests = TRUE, meanDigits = 1, sdDigits = 1, freqDigits = 2, statDigits = 2, pDigits = 5)
}

