% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{knit_print.table1}
\alias{knit_print.table1}
\title{Method for printing in a \code{knitr} context.}
\usage{
\method{knit_print}{table1}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{table1}}.}

\item{...}{Further arguments passed on to \code{knitr::knit_print}.}
}
\description{
Method for printing in a \code{knitr} context.
}
