% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1_functions.R
\name{prop_row}
\alias{prop_row}
\title{A row function}
\usage{
prop_row(y, x, margin, single)
}
\arguments{
\item{Variable}{An R object. Categorical variable of interest.}

\item{Group}{An R object. Group variable.}

\item{Margin}{a vector giving the desired margin to calculate percentages in the table for each group.
1 indicates rows, 2 indicates columns.}
}
\description{
Creates no. and percentage for each level of the variable, 
with a p-value from Chi-square test or Fisher's exact test.
}
\details{
Can handle >1 group; number of columns will adjust automatically.

Test based on expected cell value <5. Will only return row for last level of the factor.
}
\examples{
x<-rep(c('Red','Blue'),c(10,40))  
y<-rep(c('A','B'), 25) 
prop_row(x,y)
prop_row()
}
