% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{subsetp}
\alias{subsetp}
\title{Subset function that preserves column attributes.}
\usage{
subsetp(x, ..., droplevels = TRUE)
}
\arguments{
\item{x}{An object to be subsetted (usually a \code{\link{data.frame}}).}

\item{...}{Further arguments passed to \code{\link{subset}}.}

\item{droplevels}{If \code{TRUE} (the default), then unused factor levels are dropped (see \code{\link{droplevels}}).}
}
\value{
An object similar to \code{x} containing just the selected elements.
In the case of a \code{\link{data.frame}}, attributes of columns (such as
\code{\link{label}} and \code{\link{units}}) are preserved.
}
\description{
Subset function that preserves column attributes.
}
\seealso{
\code{\link{subset}}
\code{\link{droplevels}}
}
\keyword{utilities}
