% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Title}
\usage{
table1(
  dataset,
  ContVarList = NULL,
  CatVarList = NULL,
  group = NULL,
  color_list = NULL
)
}
\arguments{
\item{color_list}{}
}
\value{

}
\description{
Title
}
