% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table1}
\alias{table1}
\title{Takes a dataset and returns Table 1 with graphics for distibutions}
\usage{
table1(dataset, ContVarList, CatVarList, group = NULL, color_list = NULL)
}
\value{
table1
}
\description{
Takes a dataset and returns Table 1 with graphics for distibutions
}
